move tests to dedicated file to avoid pre-existing debug+ASAN timeouts in worker_threads.test.ts
